+++ lib/mongoid/config.rb
@@ -35,6 +35,21 @@ module Mongoid #:nodoc
     option :use_activesupport_time_zone, default: true
     option :use_utc, default: false
 
+    # keys to remove from self to not pass through to Mongo::Connection
+    PRIVATE_OPTIONS = %w(uri host port database databases username password logger)
+
+    # Get the blacklisted options from passing through to the driver.
+    #
+    # @example Get the blacklisted options.
+    #   Config.blacklisted_options
+    #
+    # @return [ Array<String> ] The blacklist.
+    #
+    # @since 2.4.7
+    def blacklisted_options
+      PRIVATE_OPTIONS + settings.keys.map(&:to_s)
+    end
+
     # Get any extra databases that have been configured.
     #
     # @example Get the extras.