+++ lib/mongoid/config/database.rb
@@ -6,9 +6,6 @@ module Mongoid #:nodoc:
     # database from options.
     class Database < Hash
 
-      # keys to remove from self to not pass through to Mongo::Connection
-      PRIVATE_OPTIONS = %w(uri database username password logger)
-
       # Configure the database connections. This will return an array
       # containing the master and an array of slaves.
       #
@@ -162,7 +159,7 @@ module Mongoid #:nodoc:
           pool_size: pool_size,
           logger: logger? ? Mongoid::Logger.new : nil,
           slave_ok: slave
-        }).merge(self).reject { |k,v| PRIVATE_OPTIONS.include? k }.
+        }).merge(self).reject { |k,v| Config.blacklisted_options.include?(k) }.
           inject({}) { |memo, (k, v)| memo[k.to_sym] = v; memo} # mongo likes symbols
       end