+++ CHANGELOG.md
@@ -484,6 +484,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#1800 Invalid options for the Mongo connection are now filtered out.
+
 * \#1785 Case equality has been fixed to handle instance checks properly.
 
 ## 2.4.6